Update tests.yml
I just removed the "Comment Coverage on PR" step in the jobs in order for GH Actions to run properly. If this is something you feel is highly desirable to have we can configure the correct permissions and put that functionality back in. I mainly just want to ensure the tests are running/passing properly. 

Fixes Sub-Issue on Issue #10 (not sure how to directly link sub-issues, but here is the actual link: https://github.com/Twiggy248/SE25-Group18-Project3/issues/16)